+++ integration-tests/src/test/java/com/transferwise/tasks/ext/management/dao/MySqlManagementTaskDaoIntTest.java
@@ -1,0 +1,4 @@
+package com.transferwise.tasks.ext.management.dao;
+
+public class MySqlManagementTaskDaoIntTest extends ManagementTaskDaoIntTest {
+}