+++ integration-tests/src/test/java/com/transferwise/tasks/ext/management/dao/PostgresManagementTaskDaoIntTest.java
@@ -1,0 +1,7 @@
+package com.transferwise.tasks.ext.management.dao;
+
+import org.springframework.test.context.ActiveProfiles;
+
+@ActiveProfiles("postgres")
+public class PostgresManagementTaskDaoIntTest extends ManagementTaskDaoIntTest {
+}