+++ integration-tests/src/test/java/com/transferwise/tasks/test/dao/PostgresTestTaskDaoIntTest.java
@@ -1,0 +1,7 @@
+package com.transferwise.tasks.test.dao;
+
+import org.springframework.test.context.ActiveProfiles;
+
+@ActiveProfiles("postgres")
+public class PostgresTestTaskDaoIntTest extends TestTaskDaoIntTest {
+}