+++ tw-tasks-core/src/main/java/com/transferwise/tasks/dao/ITaskSqlMapper.java
@@ -1,0 +1,10 @@
+package com.transferwise.tasks.dao;
+
+import java.util.UUID;
+
+public interface ITaskSqlMapper {
+
+  UUID sqlTaskIdToUuid(Object object);
+
+  Object uuidToSqlTaskId(UUID uuid);
+}