+++ tw-tasks-core/src/main/java/com/transferwise/tasks/dao/PostgresTaskSqlMapper.java
@@ -1,0 +1,16 @@
+package com.transferwise.tasks.dao;
+
+import java.util.UUID;
+
+public class PostgresTaskSqlMapper implements ITaskSqlMapper {
+
+  @Override
+  public UUID sqlTaskIdToUuid(Object object) {
+    return (UUID) object;
+  }
+
+  @Override
+  public Object uuidToSqlTaskId(UUID uuid) {
+    return uuid;
+  }
+}